refactor: transfer validation logic from controller into service

+++ src/users/service.ts
@@ -1,5 +1,8 @@
+import { validate } from "uuid";
+import ApiError from "../apiError/apiError";
 import { userRepository } from "./repository";
-import { User } from "./types";
+import { ErrorMessages } from "../apiError/constants";
+import { isUser } from "./utils";
 
 export const userService = {
   async getAll() {
@@ -7,18 +10,33 @@ export const userService = {
   },
 
   async getOne(id: string) {
+    if (!validate(id)) {
+      throw ApiError.badRequest(ErrorMessages.INVALID_ID);
+    }
     return userRepository.getOne(id);
   },
 
-  async create(user: User) {
+  async create(user: unknown) {
+    if (!isUser(user)) {
+      throw ApiError.badRequest(ErrorMessages.INVALID_DATA);
+    }
     return userRepository.create(user);
   },
 
   async delete(id: string) {
+    if (!validate(id)) {
+      throw ApiError.badRequest(ErrorMessages.INVALID_ID);
+    }
     return userRepository.delete(id);
   },
 
-  async update(id: string, user: User) {
+  async update(id: string, user: unknown) {
+    if (!validate(id)) {
+      throw ApiError.badRequest(ErrorMessages.INVALID_ID);
+    }
+    if (!isUser(user)) {
+      throw ApiError.badRequest(ErrorMessages.INVALID_DATA);
+    }
     return userRepository.update(id, user);
   },
 };
